+++ src/main/java/at/hannibal2/skyhanni/utils/TimeUtils.kt
@@ -2,6 +2,7 @@ package at.hannibal2.skyhanni.utils
 
 import at.hannibal2.skyhanni.utils.NumberUtil.addSeparators
 import at.hannibal2.skyhanni.utils.StringUtils.matchMatcher
+import io.github.moulberry.notenoughupdates.util.SkyBlockTime
 import kotlin.time.Duration
 
 object TimeUtils {
@@ -116,6 +117,21 @@ object TimeUtils {
                 throw RuntimeException("Invalid format: '$string'")
             }
         }.toLong()
+    }
+
+    fun SkyBlockTime.formatted(): String {
+        val hour = if (this.hour > 12) this.hour - 12 else this.hour
+        val timeOfDay = if (this.hour > 11) "pm" else "am" // hooray for 12-hour clocks
+        var minute = this.minute.toString()
+        if (minute.length != 2) {
+            minute = minute.padStart(2, '0')
+        }
+
+        val month = SkyBlockTime.monthName(this.month)
+        val day = this.day
+        val daySuffix = SkyBlockTime.daySuffix(day)
+        val year = this.year
+        return "$month $day$daySuffix, Year $year $hour:${minute}$timeOfDay" // Early Winter 1st Year 300, 12:03pm
     }
 }
